requirements.txt
numpy==1.22.4
matplotlib==3.5.3
open3d==0.17.0
opencv-python==4.7.0.72
Pillow==9.3.0
PyYAML==6.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
tqdm==4.66.1
yacs==0.1.8
imageio==2.21.2